syncronised as much as possible the files to those of comm-v2. removed the file EoSensors.h. use EoAnalogSensors.h instead
svn path=/trunk/iCub/firmware/emBODY/eBcode/embobj/; revision=27588

+++ emBODY/eBcode/arch-arm/embobj/plus/comm-v1/icub/EoBoards.h
@@ -42,9 +42,7 @@ extern "C" {
 // - external dependencies --------------------------------------------------------------------------------------------
 
 #include "EoCommon.h"
-#include "EoMeasures.h"
 
-#include "EoMotionControl.h" // serve?
 
 // - public #define  --------------------------------------------------------------------------------------------------
 
@@ -54,7 +52,7 @@ extern "C" {
 
 // - declaration of public user-defined types ------------------------------------------------------------------------- 
 
-typedef uint16_t eObrd_boardId_t;
+typedef uint8_t eObrd_boardId_t;
 
 
 typedef enum
